chore: move gate to common
