+++ packages/waypoint/web/auth.ts
@@ -1,10 +1,10 @@
 import { Address } from "viem"
 
 import { CommunicateHelper } from "../common/communicate"
+import { RONIN_WAYPOINT_ORIGIN_PROD } from "../common/gate"
 import { openPopup, replaceUrl } from "../common/popup"
-import { RONIN_WAYPOINT_ORIGIN_PROD } from "./common/gate"
-import { WaypointResponse } from "./common/id-response"
 import { getScopesParams, Scope } from "./common/scope"
+import { WaypointResponse } from "./common/waypoint-response"
 import { validateIdAddress } from "./utils/validate-address"
 
 /**
