+++ packages/waypoint/web/common/waypoint-response.ts
