+++ packages/waypoint/web/provider.ts
@@ -12,13 +12,13 @@ import {
 } from "viem"
 
 import { CommunicateHelper } from "../common/communicate"
+import { RONIN_WAYPOINT_ORIGIN_PROD } from "../common/gate"
 import { openPopup } from "../common/popup"
 import { BaseAuthorizeOpts } from "./auth"
 import { VIEM_CHAIN_MAPPING } from "./common/chain"
 import { Eip1193EventName, Eip1193Provider, RoninWaypointRequestSchema } from "./common/eip1193"
-import { RONIN_WAYPOINT_ORIGIN_PROD } from "./common/gate"
-import { WaypointResponse } from "./common/id-response"
 import { getScopesParams, Scope } from "./common/scope"
+import { WaypointResponse } from "./common/waypoint-response"
 import { personalSign } from "./core/personal-sign"
 import { sendTransaction } from "./core/send-tx"
 import { signTypedDataV4 } from "./core/sign-data"
